Sanity-check streaming duplex operations against bulk ones
For pseudorandom bulk data, check that a single padded absorb, decrypt
or encrypt has the same effect on the sponge state and data as the same
operation performed in chunks of varying size.

Similarly, check that a single large squeeze operation is equivalent to
squeezing the same amount of data chunk-by-chunk.