Add X25519 test vectors from RFC 7748 and Project Wycheproof
The published outputs assume RFC 7748 bit-clamping of both scalars and
points, so we implement clamp() and use it before each x25519() call. A
discussion of clamping and non-contributory behaviour will be needed in
the documentation for this library.

Wycheproof includes values labelled as from RFC 7748, but these are
different to those listed in the RFC itself. For completeness, the
original versions are prepended to the list to check. The second RFC
test vector is interesting as it checks for clamping of the top-bit of
the public key as well as the low and high bits of the scalar.

x25519() returns -1 where the resulting shared key is zero, to flag
non-contributory behaviour. Values intended to trigger this are marked
as ZeroSharedSecret in the Wycheproof x25519_test.json. We check that
these values and only these values result in a non-zero x25519() return.

x25519.c passes all of the test vectors, including those crafted to
trigger edge-cases in carry propagation or the Montgomery ladder.