Simplify the gimli.h streaming API
Allow duplex calls to seamlessly absorb/squeeze/encrypt/decrypt partial
chunks in a streaming setting by accepting, updating and returning a
rate offset as well as the Gimli state itself.

Previous attempts at this imposed an unacceptable performance penalty of
10-20% on x86-64 and arm64, but explicitly separating out the head, main
body and tail of each operation appears to be sufficient for both gcc
and clang to match the high speed of the original implementation.