Expand out the power chain in x25519()
To invert z2, the current x25519() power chain interprets a set of
square-and-multiply steps defined in a constant struct array, aiming to
minimise object code for very constrained embedded platforms.

Add a mulsqrn() function as a shorthand for the squaring loop, and write
out the power chain as an explicit sequence of calls to this.

This slightly shortens and speeds up the code at the cost of slightly
larger object code, but more importantly, it makes the code less
mysterious and easier to read - a better trade-off for Pocketcrypt.