Implement Elligator 2 forward mapping for X25519
Add invsqrt() to compute the inverse square root of a field element,
using a square-and-multiply chain to raise to the power (p - 5)/8, i.e.
2**252 - 3. Return 0 or -1 depending on the quadratic character of the
element in the field, suitable for use as a condswap() mask.

Use this to efficiently implement the Elligator 2 forward mapping from
https://elligator.cr.yp.to/elligator-20130828.pdf as x25519_point(),
choosing non-square parameter u = 2 which is suggested for curve25519.

Add documentation and tests against known images generated with other
implementations, including random elements and specific edge-cases.