Substitute a more pragmatic check for vector extensions
The correct way to detect vector extensions is with __has_builtin(), but
this requires a much newer compiler than the extensions themselves need.
For example, __has_builtin() is not supported before gcc 10.1.0, but
vector extensions work fine as far back as gcc 5.1.0.

Replace the __has_builtin() feature test with a more pragmatic check:
first for clang >= 4.0.0 where __builtin_shufflevector() is available,
then for gcc 5.x compatibility implying support for __builtin_shuffle().

Apple report clang version numbers incorrectly in Xcode, but the test
for __clang_major__ >= 4 is a reasonable one in their sequence too.

Include advice about the supported versions of gcc and clang in the
error message from duplex.h when vector extensions are not detected.