Rewrite duplex operations in explicit vector form
Although the duplex operations benchmark very well, the implementation
relies heavily on the compiler to vectorise the XOR loops. Sometimes
when the functions are inlined in larger programs, and particularly with
older versions of gcc/clang, optimisation opportunities can be missed.
Performance then drops off by 15% because of byte-wise memory access.

Rewrite the duplex loops with a byte-wise loop for initial and final
partial chunks, but explicit vector code for the body of the operation.

We use __builtin_memcpy() to load from and store to data buffers as this
efficiently handles misaligned access, unlike a cast and dereference
which will be invalid on some platforms. There doesn't appear to be any
performance penalty from using it in the optimal aligned case.