Integrate the counter into the duplex_t state
Extend the duplex_t array from uint32x4_t[3] to uint32x4_t[4], using the
extra space to include the uint64_t counter as part of the duplex state.

This simplifies duplex calls and reduces the likelihood of padding
errors from mishandling a separate counter. For example,

  duplex_t state = { 0 };
  size_t counter = 0;
  counter = duplex_absorb(state, counter, key, sizeof(key));
  counter = duplex_absorb(state, counter, nonce, duplex_rate);
  counter = duplex_encrypt(state, counter, message, length));
  counter = duplex_pad(state, counter);
  counter = duplex_squeeze(state, counter, tag, duplex_rate);

can now be rewritten more succinctly as

  duplex_t state = { 0 };
  duplex_absorb(state, key, sizeof(key));
  duplex_absorb(state, nonce, duplex_rate);
  duplex_encrypt(state, message, length));
  duplex_pad(state);
  duplex_squeeze(state, tag, duplex_rate);

This change has no measurable impact on x86-64 or arm64 benchmarks with
gcc and clang. (An earlier patch wrapping the state and counter in a
duplex_t struct was abandoned because it adversely affected performance
on clang builds.)