Add simple demonstration tools
Provide some examples to illustrate how to use Pocketcrypt to construct
simple cryptographic building-blocks like key generation, authenticated
encryption and signatures. The build script can compile these programs
in-place or install them outside the tree.

For simplicity, we assume that the kernel will not leak memory contents
after a process exits, and that getentropy() is available and secure.