Rework the duplex_swirl() memory-hard example
Allow the data-independent and data-dependent round counts to be
specified independently. If no data-independent rounds are requested,
the seed is not accessed and can be NULL.

Typecast the buffer as uint32x4 cells[pages][64] rather than the
equivalent uint32x4 cells[pages << 6] to clarify the dependency graph
in the subsequent code.

XOR both back-references into the sponge when filling each page rather
than relying on the implicit dependency on the previous page via the
single-threaded sponge. This may be somewhat redundant, but it has no
performance impact, makes the graph explicit, and ensures that pebbling
analyses of Argon2B from the literature (modelling its stateless hash
as a random oracle) work for this construction unmodified. Like
Argon2B, we rely on theorem 7 of https://eprint.iacr.org/2018/944.pdf
to show that XORing oracle inputs instead of concatenating is fine
provided each page has unique parents.

To change out the Argon2B DAG for (say) DRSample, it is sufficient to
adjust the definition of the non-uniformly distributed 32-bit offset in
terms of the uniformly distributed 32-bit key.

Note that duplex_swirl() still assumes it is called with rate-aligned
secret and seed states. Use duplex_pad() beforehand if this is not
already guaranteed by the protocol.