Fix code style in x25519.c
Lower-case the LIMB() macro to match the style in duplex.h.

LIMBS and WBITS need not be macros at all, but LIMBS is used in array
declarations so needs to be an integer constant expression to avoid
variable length arrays. Use an anonymous enum to create 'sufficiently
constant' identifiers for limbs and width, given that C99 and C11 have
decided that const ints aren't integer constants.