Demonstrate chunked authenticated encryption in the examples
In real applications, it is usually better to implement bulk encryption
with chained authenticated chunks rather than a single tag at the end
of the stream. This sidesteps the need to buffer the entire stream and
check the final tag before any plaintext can safely be released.

https://eprint.iacr.org/2013/772.pdf more formally describes the
resulting mode in a duplex context, calling it 'MAC and continue'. A
tag is squeezed at the end of the ciphertext for each (padded) chunk,
before proceeding to the next. The tag authenticates the entire
preceding transcript, i.e. the whole stream up to and including the
tagged chunk.

Provide more realistic examples by implementing this mode. In practice,
it adds no complexity, even simplifying decryption by removing the need
to buffer arbitrary-length streams.

Take advantage of the resulting opportunity to eliminate heap allocation
in favour of simpler static arrays wherever possible.